refactor: add separate API service method for updating SCIM server
